+++ aws/init.go
@@ -32,7 +32,7 @@ import (
 )
 
 var (
-	AccessService, InfraService, StorageService, NotificationService, QueueService, DnsService, LambdaService, MonitoringService, CdnService cloud.Service
+	AccessService, InfraService, StorageService, NotificationService, QueueService, DnsService, LambdaService, MonitoringService, CdnService, CloudformationService cloud.Service
 )
 
 func InitServices(conf map[string]interface{}, log *logger.Logger) error {
@@ -56,6 +56,7 @@ func InitServices(conf map[string]interface{}, log *logger.Logger) error {
 	LambdaService = NewLambda(sess, awsconf, log)
 	MonitoringService = NewMonitoring(sess, awsconf, log)
 	CdnService = NewCdn(sess, awsconf, log)
+	CloudformationService = NewCloudformation(sess, awsconf, log)
 
 	cloud.ServiceRegistry[InfraService.Name()] = InfraService
 	cloud.ServiceRegistry[AccessService.Name()] = AccessService
@@ -66,6 +67,7 @@ func InitServices(conf map[string]interface{}, log *logger.Logger) error {
 	cloud.ServiceRegistry[LambdaService.Name()] = LambdaService
 	cloud.ServiceRegistry[MonitoringService.Name()] = MonitoringService
 	cloud.ServiceRegistry[CdnService.Name()] = CdnService
+	cloud.ServiceRegistry[CloudformationService.Name()] = CloudformationService
 
 	return nil
 }
@@ -99,6 +101,7 @@ func NewDriver(region, profile string, log ...*logger.Logger) (driver.Driver, er
 	drivers = append(drivers, NewLambda(sess, awsconf, drivLog).Drivers()...)
 	drivers = append(drivers, NewMonitoring(sess, awsconf, drivLog).Drivers()...)
 	drivers = append(drivers, NewCdn(sess, awsconf, drivLog).Drivers()...)
+	drivers = append(drivers, NewCloudformation(sess, awsconf, drivLog).Drivers()...)
 
 	return driver.NewMultiDriver(drivers...), nil
 }