+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Set/update buckets or s3objects predefined ACL (private / public-read / public-read-write / bucket-owner-read...): `awless update s3object acl=public-read`
 - List CloudFront distributions: `awless list distributions`
 - Create/Update/Check/Delete a CloudFront distribution: `awless create/update/check/delete distribution`
+- List CloudFormation stacks: `awless list stacks`
 - `awless log --raw-json` shows the full info stored on template execution (context, fillers used, region, ...). Typically this contextual info can be reused for replay and updates of templates
 
 ## v0.0.24 [2017-05-22]